requirements.txt
flask
flask-mysqldb
flask-ldap
wtforms
passlib
jira
celery
redis
pyVmomi >= 6.7
suds ; python_version < '3'
suds-jurko ; python_version >= '3.0'
tabulate
vapi-client-bindings == 1.3.1
vmc-client-bindings == 1.2.0
vapi-vmc-client
nsx-policy-python-sdk
vmc-app-python-sdk